:recycle: Split up backend file
